Fix #65 Moved AWS to a separate folder Corrected platform msg & location Made sure all test cases pass
