+++ securetea/__init__.py
@@ -6,6 +6,8 @@ from .lib.notifs import secureTeaTwitter
 from .lib.notifs import secureTeaTwilio
 from .lib.notifs import secureTeaTelegram
 from .lib.notifs import secureTeaSlack
+from .lib.notifs.aws import secureTeaAwsSES
+from .lib.notifs.aws import helper_email
 from .lib.firewall import engine
 from .lib.firewall import packet_filter
 from .lib.firewall import mapping
