+++ securetea/lib/__init__.py
@@ -3,6 +3,8 @@ from .notifs import secureTeaTwitter
 from .notifs import secureTeaTwilio
 from .notifs import secureTeaTelegram
 from .notifs import secureTeaSlack
+from .notifs.aws import secureTeaAwsSES
+from .notifs.aws import helper_email
 from .firewall import engine
 from .firewall import packet_filter
 from .firewall import secureTeaFirewall
