+++ securetea/lib/notifs/__init__.py
@@ -3,3 +3,5 @@ from . import secureTeaTwitter
 from . import secureTeaSlack
 from . import secureTeaTwilio
 from . import secureTeaTelegram
+from .aws import secureTeaAwsSES
+from .aws import helper_email
