+++ securetea/lib/notifs/secureTeaTelegram.py
@@ -51,7 +51,8 @@ class SecureTeaTelegram():
             msg (TYPE): Description
         """
         message = (str(msg) + " at " + common.getdatetime() +
-                   " " + common.get_current_location())
+                   " " + common.get_current_location() + common.get_platform())
+
         bot = telegram.Bot(token=self.token)
         bot.send_message(chat_id=self.user_id, text=message)
         self.logger.log(
